// backend/interfaces/repository.interface.ts
export interface Owner {
    login: string | undefined;
    id: number | undefined;
    node_id: string | undefined;
    avatar_url: string | undefined;
    gravatar_id: string | undefined;
    url: string | undefined;
    html_url: string | undefined;
    followers_url: string | undefined;
    following_url: string | undefined;
    gists_url: string | undefined;
    starred_url: string | undefined;
    subscriptions_url: string | undefined;
    organizations_url: string | undefined;
    repos_url: string | undefined;
    events_url: string | undefined;
    received_events_url: string | undefined;
    type: string | undefined;
    site_admin: boolean | undefined;
}

export interface License {
    key: string | undefined;
    name: string | undefined;
    spdx_id: string | undefined;
    url: string | undefined;
    node_id: string | undefined;
}

export interface RepositoryInterface {
    id: number | undefined;
    node_id: string | undefined;
    name: string | undefined;
    full_name: string | undefined;
    private: boolean | undefined;
    owner: Owner | undefined;
    html_url: string | undefined;
    description: string | undefined;
    fork: boolean | undefined;
    url: string | undefined;
    forks_url: string | undefined;
    keys_url: string | undefined;
    collaborators_url: string | undefined;
    teams_url: string | undefined;
    hooks_url: string | undefined;
    issue_events_url: string | undefined;
    events_url: string | undefined;
    assignees_url: string | undefined;
    branches_url: string | undefined;
    tags_url: string | undefined;
    blobs_url: string | undefined;
    git_tags_url: string | undefined;
    git_refs_url: string | undefined;
    trees_url: string | undefined;
    statuses_url: string | undefined;
    languages_url: string | undefined;
    stargazers_url: string | undefined;
    contributors_url: string | undefined;
    subscribers_url: string | undefined;
    subscription_url: string | undefined;
    commits_url: string | undefined;
    git_commits_url: string | undefined;
    comments_url: string | undefined;
    issue_comment_url: string | undefined;
    contents_url: string | undefined;
    compare_url: string | undefined;
    merges_url: string | undefined;
    archive_url: string | undefined | undefined;
    downloads_url: string | undefined;
    issues_url: string | undefined;
    pulls_url: string | undefined;
    milestones_url: string | undefined;
    notifications_url: string | undefined;
    labels_url: string | undefined;
    releases_url: string | undefined;
    deployments_url: string | undefined;
    created_at: Date | undefined;
    updated_at: Date | undefined;
    pushed_at: Date | undefined;
    git_url: string | undefined;
    ssh_url: string | undefined;
    clone_url: string | undefined;
    svn_url: string | undefined;
    homepage: string | undefined;
    size: number | undefined;
    stargazers_count: number | undefined;
    watchers_count: number | undefined;
    language: string | undefined;
    has_issues: boolean | undefined;
    has_projects: boolean | undefined;
    has_downloads: boolean | undefined;
    has_wiki: boolean | undefined;
    has_pages: boolean | undefined;
    forks_count: number | undefined;
    mirror_url?: any;
    archived: boolean | undefined;
    disabled: boolean | undefined;
    open_issues_count: number | undefined;
    license: License | undefined;
    forks: number | undefined;
    open_issues: number | undefined;
    watchers: number | undefined;
    default_branch: string | undefined;
}